Changed layout, added analytics to home page and added role protection to controllers

+++ GenericStockManagement/Controllers/CategoryController.cs
@@ -16,16 +16,41 @@ namespace GenericStockManagement.Controllers
         }
         public IActionResult Index()
         {
-            return View();
-        }
+            var products = _productRepository.GetAll();
+            var categories = _categoryRepo.GetAll();
 
+            var model = new HomeViewModel(products, categories);
+
+
+            return View(model);
+        }
+        public IActionResult CategoryActions(List<int> selectedCatIds, string submitButton)
+        {
+            if (submitButton == "Delete")
+            {
+                return RedirectToAction("Delete", "Category", new { ids = selectedCatIds });
+            }
+            else if (submitButton == "Update")
+            {
+                return RedirectToAction("Update", "Category", new { ids = selectedCatIds });
+            }
+            return RedirectToAction("Index");
+        }
         public IActionResult Create()
         {
+            if (HttpContext.Session.GetString("Role") != "Admin" && HttpContext.Session.GetString("Role") != "Contributor")
+            {
+                return RedirectToAction("Login", "Account");
+            }
             return View();
         }
         [HttpPost]
         public IActionResult Create(Category entity)
         {
+            if (HttpContext.Session.GetString("Role") != "Admin" && HttpContext.Session.GetString("Role") != "Contributor")
+            {
+                return RedirectToAction("Login", "Account");
+            }
             var newCategory = new Category()
             {
                 Name = entity.Name,
@@ -35,7 +60,7 @@ namespace GenericStockManagement.Controllers
             if (ModelState.IsValid)
             {
                 _categoryRepo.Add(newCategory);
-                return RedirectToAction("Index", "Home");
+                return RedirectToAction("Index", "Category");
             }
 
             return View(newCategory);
@@ -43,6 +68,10 @@ namespace GenericStockManagement.Controllers
 
         public IActionResult Delete(List<int> ids)
         {
+            if (HttpContext.Session.GetString("Role") != "Admin")
+            {
+                return RedirectToAction("Login", "Account");
+            }
             var categories = new List<Category>();
             var categoryProducts = new Dictionary<int, List<Product>>();
 
@@ -64,15 +93,23 @@ namespace GenericStockManagement.Controllers
         [HttpPost]
         public IActionResult ConfirmDelete(List<int> selectedIds)
         {
+            if (HttpContext.Session.GetString("Role") != "Admin")
+            {
+                return RedirectToAction("Login", "Account");
+            }
             foreach (var id in selectedIds)
             {
                 _categoryRepo.Delete(id);
             }
-            return RedirectToAction("Index", "Home");
+            return RedirectToAction("Index", "Category");
         }
 
         public IActionResult Update(List<int> ids)
         {
+            if (HttpContext.Session.GetString("Role") != "Admin" && HttpContext.Session.GetString("Role") != "Contributor")
+            {
+                return RedirectToAction("Login", "Account");
+            }
             var categories = ids.Select(id => _categoryRepo.GetById(id)).ToList();
             var viewModel = new CategoryListViewModel(categories, null);
             return View(viewModel);
@@ -81,11 +118,15 @@ namespace GenericStockManagement.Controllers
         [HttpPost]
         public IActionResult ConfirmUpdate(List<Category> updatedCategories)
         {
+            if (HttpContext.Session.GetString("Role") != "Admin" && HttpContext.Session.GetString("Role") != "Contributor")
+            {
+                return RedirectToAction("Login", "Account");
+            }
             foreach (var category in updatedCategories)
             {
                 _categoryRepo.Update(category);
             }
-            return RedirectToAction("Index", "Home");
+            return RedirectToAction("Index", "Category");
         }
     }
 }
